Typo fix in Bitcoin Package
Fixed references to Ripple and Ethereum that did not seem to fit Bitcoin's branch. Please revert if incorrect.

Also, should Bitcoin address creation be referred to as account creation, despite it not being an account based blockchain? I could see why it would be useful to keep the "account" moniker across multiple chains for simplicity.

-GSM